Adding override Thyra::DefaultSpmdMultiVector::contigSubViewImpl(...) (Bug 4818)
After three lovely hours debugging a failing Salinas/Anasazi/Thyra test, we
found out that a recent refactoring in Anasazi to correctly create const views
exposed a problem with Thyra::DefaultSpmdMultiVector.  The problem is that the
override for the contigSubViewImpl(...) function was missing which was
therefore using the default contigSubViewImpl(...) function that creates a
Thyra::DefaultColumnwiseMultiVector object.  The Salinas Thyra::LinearOpBase
subclass did not like this very much since it was (rightly) expecting to get a
Thyra:::SpmdMultiVectorBase object and a dynamic cast failed.

The fix is to just add a trivial override of the contigSubViewImpl(...)
function.  I have added a unit test exposing the problem and then added the
code to fix the problem.   The current unit tests for
Thyra::DefaultSpmdMutiMultiVector do a pretty good job of pinning down the
behavior with respect to contiguous and non-contiguous subviews.

=============================
Build/Test Cases Summary
=============================

Fri Jun 11 18:03:17 MDT 2010

Enabled Packages: Thyra
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=382,notpassed=0 (21.1800458153 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=407,notpassed=0 (20.7825719158 min)

Local commits for this build/test group:
----------------------------------------
1d0b171 Adding override Thyra::DefaultSpmdMultiVector::contigSubViewImpl(...) (Bug 4818)
60ede6c Adding THYRA_DEPRECATED