Replace existing Makefile export system with one based off of find Trilinos
The old Makefile export system while functional had several issues with it.
Since a lot of the required functionality was shared with the find Trilinos
capability it the Makefile export system was re-done to take advantage of that
work. Right now both the Makefile export and find Trilinos share all the logic
for determining which packages/libraries/includes/TPLs/etc were enabled and the
proper ordering for correct linking. This way all the functionality for one can
be guaranteed for both. The only difference is that the compiler decorations for
adding include/library paths and libraries are added to the appropriate
variables so that they can be directly used by a Makefile. For example what
would appear in a TrilinosConfig.cmake file as "Epetra;Epetraext" would appear
in the Makefile.export.Trilinos as "-lEpetra -lEpetraext".

There are some importants changes from the old Makefile Export system. The most
important is that the variable names have been changed. Since many new variables
were added compared to the old Makefile export since there were a lot more
variables in the find Trilinos set that could be useful in Makefiles as well.
Because of that all the variables in the Makefile export files are named exactly
the same as they were in the find Trilinos capability, which means that package
names are not all uppercase. Also compared to the old autotools system
specifically many variables that were previously one variable have been split
out into several variables, for instance a package's dependent libraries and the
TPL libraries that it uses are now two different variables.  This will likely
mean all projects which currently use the curent Makefile export or the old
autotools version will have to be updated to use the new variables.

documentation and a simple example for this capability and the find Trilinos
capability has been moved to commontools/importing.

=============================
Build/Test Cases Summary
=============================

Wed Jun 16 14:24:33 MDT 2010

Enabled Packages: TrilinosFramework
Enabled all Packages

Build test results:
-------------------
0) MPI_DEBUG => FAILED: Trilinos/MPI_DEBUG: passed=1204,notpassed=1 => Not ready for final commit/push! (0.0 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=1253,notpassed=0 (0.0 min)

Local commits for this build/test group:
----------------------------------------
8986a57 Replace existing Makefile export system with one based off of find Trilinos

***
*** WARNING: The acceptance criteria for doing a commit/push has *not*
*** been met, but a commit/push is being forced anyway by --force-commit-push!
***