Adding a plain-source copy of eg to avoid having to use install-git.py on most computers
Here I have added a plain-source copy of eg so that on many machines that have
a fairly recent version of git, they can just clone Trilinos directly with git
and then update their path to include this eg script and then be able to
directly use eg without having to run the full install-git.py script.  That
does not seem like a lot of trouble for people to do.