Another fix to the Makefile.export capability
This commit addresses bug 4851.

Kevin Long noticed that the rpath was not being set for shared builds. The
variable was there, but the logic to populate it was missing. It is now being
set for both install and build tree versions, however, I'm not 100% sure that
the build tree version is portable. The implementation of rpath is not very
standard across compilers I know. I put in an implementation that works for the
compilers I have used rpath on, and I believe is reasonably common, but I can't
guarantee that it will work for every compiler. The installation version is
safer since it only has one path it is unlikely to cause issues so long as the
"rpath option" from cmake is correct for the machine. It happened to not be set
for mac platforms so it is possible that it is wrong on others as well, but is
outside the scope of these changes.

I also did some clean up, I noticed that the ld_flags variable was duplicated in
the package configure files for some reason so I pruned one of them and then
made sure that the trilinos configure files were consistent with the package
files.

There were a couple places were -l and -L were hard coded, there are cmake
variables that hold the compiler specific options so I changed those cases to
use them.

The MPI utility variables were not being set for the package makefile.export
files. This shouldn't have been a problem for anyone since that information
would be set in the tpl library/include variables, but I fixed it just to be
consistent.