more fixes for Trios configure problems
Added netcdf to the s903186 builds. I missed this system yesterday.

Also added the netcdf paths for the continuous testing on trilinos-test2. When I
checked this system yesterday netcdf was installed and appeared to be set up so
I figured it was done. It turns out that it was only set up for the nightly
builds and the continuous testing couldn't find netcdf. I have updated the core
file so that all testing on trilinos-test2 through these scripts will find
netcdf properly now.