Adding a debugging feature that's turned on when debug level >= 1
A customer is running a very large problem which encounters a
ZOLTAN_PRINT_ERROR() on just one process.  It would help to see
a back trace at this point.  This commit causes a back trace
(generated from ZOLTAN_TRACE_ENTER/ZOLTAN_TRACE_EXIT calls) to
be printed by the process that hits the ZOLTAN_PRINT_ERROR().
I believe that asking the customer to turn on traces for all
processes would not be practical given the size of the problem.

This back trace is fragile due to the fact that it would be
wrong if the ENTER/EXIT calls are not perfectly matched in the code.