Adding Zoltan_overflow_test
64 bit zoltan assumes that the number of rows or column in a
process group (dist_x[k] - dist_x[k-1]) fits in an "int".
Zoltan_overflow_test checks that the value of a particular "size_t"
fits in an "int".

In phg_build.c we exit with an error if overflow occurs.  An
error message suggests the number of processes necessary to
solve the problem.