Fixed bug #4903 in which repeated inserts to a CrsGraph row, requiring new allocation, produces incorrect behavior.
This bug is seen only on CPU platforms. It results from a failure to recognize
that the buffers and their views are identical, so that view is inadvertantly
changed by changing the reference. This is not present in dev/10.6 due to
changes; upon documentation, the bug was easily discovered and fixed. A patch has been pushed to the reporter; others will get the patch in the next 10.4.x release.

Build/Test Cases Summary
Enabled Packages: Tpetra
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=539,notpassed=0 (49.08 min)
1) SERIAL_RELEASE => passed: passed=533,notpassed=0 (38.54 min)