Fixing the behavior of the checkin-test.py script when ammending commit message with existing test block (bug 4861)
I finally got bit by this bug where if you run the checkin-test.py script on a
commit that already has a commit message, then the script would bomb.  The
problem was that when I refactored the script to reduce the amount of bioler
plate output, I never updated the unit tests for the new format.

Now I updated the unit tests for the current format and showed that the
behavior was broken.  I then fixed the python code to amend the commit message
correctly and it works very nice now.