Disable a couple Teuchos tests in development mode
The Refcount pointer tests need to be disabled in development mode as they rely
on files that are not in the Teuchos tree and are excluded from a tarball which
prevents teuchos from configuring. Since there should be very little development
on Teuchos on the release branch and the tests in question are for ensuring
documentation is up to date with the code I don't think there is much risk
envolved here.