(From Sierra) Parser: Keywords may now contain parameters within
This represents a signification change to how line commands are
determined.  Aria adopted an approach which placed non-keyword data
within the bounds of the keywords.  For example:

   BC FLUX FOR ENERGY on Surface_1 = CONSTANT

 * "BC FLUX FOR ENERGY" and "= CONSTANT" are the keywords
 * on is a token
 * Surface_1 is a parameter

This required a change in how a keyword is discovered and how
parameters are processed.  To process a line command, the entire
command is tested against the command's specifications.
Unfortunately, some command keywords are subsets of each other and as
a results multiple candidates are found.  For the time being, the
longest matching keyword is selected.

A follow-on task will add a method to specify the end of a split
keyword.

Build/Test Cases Summary
Enabled Packages: STK
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=4,notpassed=0 (1.25 min)
1) SERIAL_RELEASE => passed: passed=4,notpassed=0 (0.90 min)