(From Sierra) Trilinos changes to enable use of Trilinos/Zoltan in SIERRA
There are a few files in the Trilinos repository that are used to couple
Trilinos into SIERRA.  This is a separate commit for just those files.

NOTE: Deleting the old Zoltan source files will not delete object and library files associated
with Zoltan 2.01.  Care must be taken to remove the old libraries to avoid linking two
versions of Zoltan.  If using the bake command the --super-clean option should do it.