Added PyTrilinos class Epetra_NumPySerialSymDenseMatrix
The Epetra_SerialSymDenseMatrix class is in the category of classes
singled out for special treatment as a hybrid Epetra/NumPy class in
PyTrilinos, because it is a wrapper around contiguous (strided)
homogeneous data.  This header and implementation file define to C++
class that defines the hybrid constructors and other methods.