Fixed a problem that can be seen with mpl::apply for OrthogPoly.
Some compilers incorrectly match Sacado::PCE::OrthogPoly as a template
with one template parameter, even though it has two (because the second
has a default value), which can cause an ambiguous template
instantiation when mpl::apply is used.  I removed the default template
parameter, and users will just have to explicitly specify the storage.
In most cases, this is a 1-line code change.