Fixing graphviz output for boost versions greater than 1.43
Starting with boost v 1.44 there were changes to the graphviz reader/writer.  A number of typedefs of various graph types were removed.  This broke the graphviz writer in phalanx.  This update fixes the reader/writer.  The new boost code is not compatible with older versions of boost.  So we have an ifdef that uses the old code if using boost v 1.42 and 1.43 and uses the new code if 1.44 or greater.

Added test problem for new graphviz functions.