(From Sierra) IOSS: Reorganize source structure; decouple from Sierra
Reorganizing the source structure as we try to disentagle code from
the framework hierarchy.  The io_system used to contain just ioss
source code, but will now also contain interface and bridge routines
that do not depend on framework. To make this more clear; an "ioss"
subdirectory has been added to io_system and a fem_io subdirectory
added at the same level.  The ioss subdirectory is a symbolic link
into the Trilinos Trios package; this almost looks as it did before
except that there can now be Trilinos and non-Trilinos resident code
in the io_system.  ioss is in Trilinos and fem_io is (currently) not
in Trilinos, but will probably be moved there after it stabalizes.

The ioss/io_interface code was moved to the io_system/fem_io/interface
location which removes it from Trilinos for awhile.

Part of eliminating Sierra dependencies involves eliminating use of
sierra::String in favor of std::string.  This was done for some of the
Frio classes which will be moved over to fem_io in a later commit.