Upgraded PyTrilinos test scripts from getopt to optparse
Several years back I changed some scripts from the more modern
optparse module for command-line argument parsing to the older getopt
module.  There was evidence that the module was causing memory
allocation errors.  However, that evidence turned out to be false: the
problem was a mismatch of C++ and Fortran compiler versions.  With
this commit, I have re-upgraded all of the PyTrilinos test scripts
that used getopt to optparse.

In response to Bug 5000.