Adding --no-rebase option to checkin-test.py (bug 5022)
Here I added the option --no-rebase to skip the final rebase before the final
amend and push.  This will now allow the checkin-test.py script to be used
with more complex branching workflows but see ee bug 5022 for the full
justification for doing this.  I added a great unit test to test the new
bahavior and the old behavior is well maintained as well.  I did not test this
on a real set of commits but given this is a new option that is off by default
I can 99.99% guarantee that this will not break any existing use cases.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=81,notpassed=0 (0.82 min)
1) SERIAL_RELEASE => passed: passed=81,notpassed=0 (0.79 min)