fix so tests of different configuratinos get the same answers
Zoltan_Hash now returns the same hash value for 64 bit input
that is does for 32 bit input.  It does this by only hashing
the low order 32 bits.  For consecutive spaces of global IDs,
this should give equally spaced out hash values.

This is so that tests with 64 bit IDs get the same answer as
tests with 32 bit IDs.

If there is another use case where this is bad, maybe we need
to have two Zoltan_Hash functions.  The new one would not
be used when hashing global IDs.