Adding -fno-strict-aliasing only when compiling Teuchos_stacktrace.cpp with g++
The file Teuchos_stacktrace.cpp generates "warning: dereferencing type-punned
pointer will break strict-aliasing rules" which is elevated to an error
because of -Werror.  With older versions GCC than 4.2, you can't selectively
disable warnings in the code so this is the only way that I can compile this
file successfully with g++.  I verified that this is the only file that gets
-fno-strict-aliasing when compiling.

Thank goodness that CMake allows you to set per-file compiler options like
this our this code would be a no-go.