Updated PyTrilinos for new Teuchos XML and ParameterList interfaces
Recent updates to the Teuchos interfaes for ParameterLists and
XMLObjects broke the PyTrilinos interface for Teuchos.  Specifically,
the XMLObject methods addAttribute and getWithDefault changed from
being non-templated to being templated, and the classes Dependency and
DependencySheet were added.

The new templated methods resulted in no concrete instances of those
methods being added to the PyTrilinos interface.  These changes re-add
a single addAttribute and getWithDefault method to the XMLObject
class, and are written to take advantage of the untyped nature of
python arguments.

I currently choose to ignore the new Dependency and DependencySheet
classes, meaning I have to %ignore the methods in
XMLParameterListReader and XMLParameterListWriter that reference these
new classes, and provide methods that treat these arguments with their
default values.