Adding install-gcc.py to get, build, and install GCC 4.5.1 from source.
I have tested this on my RedHat Linux Enterprise 5 computer and it works fine
to build Trilinos.  However, I can't build shared libraries of Trilinos with
it so I need to look into this some more.  Or, I might just wait for the next
release of GCC 4.5.x and try that.