MueLu Aggregation code
Awesome ! My driver do something.
It creates a matrix and calls the aggregation routine on a Epetra CrsMatrix.
The driver mimics what was done by Ray to call the routine in ML (see ML_Linker for code relics) but there is no more ML specific code. The code directly works on an Epetra CrsMatrix.

Is the aggregates correct ? Hard to say. But seems OK to me.