Fixed apply() is changed to ApplyInverse() for user-defined prec's.
The code that was here wrapped the user-defined preconditioner in an
Epetra_InvOperator, which translated Apply() to ApplyInverse().  There
is a simpler way to do this when the operator is wrapped in Thyra
however, the builder function taken an enum to determine whether apply()
calls Apply() or ApplyInverse().  I changed it to use that instead.  The
main reason for doing this is it prevents stratimikos from wrapping the
preconditioner with yet an other operator to translate back to
ApplyInverse().  Most notably, this fixes the Label() function so that
the proper preconditioner label is printed when you call AztecOO,
instead of just "User defined preconditioner".