Converting Teuchos dependency on BinUtils to a proper TPL (bug 5039)
Here I am converting the new optional Teuchos dependence on the BinUtils
library to be a proper TPL.  However, this TPL is tentatively enabled by
default so that it will be enabled whenever possible.  This will result in a
little extra output from the CMake configure and on systems were BinUtils
cannot be present this will likely result in some annoying output.  However, I
believe that this is the right thing to do and we should tentatively enable a
lot more TPLs like Boost, PThreads and many other TPLs.