MueLu::SaPFactory.Build() much more like MueMat now.
The flow control of Build is very similar to that of MueMat, version 12/15/2010.
In general, Build will throw exceptions, with helpful messages, if you access a feature
that's not implemented yet.  Here are the two cases that aren't implemented yet:
    reUsePtent_ but Ptent or Coarse Nullspace don't exist
    no fine grid nullspace saved

Other changes:

Initializing damping factor to 4/3.
Max eigenvalue hardwired to 2 for now (1D constant coefficient Poisson).
Fixed MueLu_UseShortNames.hpp and SaPFactory now that MueLu::Utils is a class.
Fixing misspelling of macro in Cthulhu_MultiVectorFactory.hpp
  CTHULHU_MULTIVECTOR_FACTORY_SHORT ==> CTHULHU_MULTIVECTORFACTORY_SHORT