Implemented triple-matrix product to form coarse grid operator.
This involved implementing true transpose of prolongator in TransPFactory and
a triple matrix produce in RAPFactory.

comments
--------
This only works for underlying Epetra_CrsMatrix types.  It essentially uses EpetraExt's
capability to calculate C = A^T * B.  In our case, A=P and B=I, so C=R.  Along the way,
added transpose argument to Utils::TwoMatrixMultiply, which lets us pass the transpose
on to EpetraExt.  Also added not-implemented exception if you are using Tpetra and try
to build coarse A.

In TwoMatrixAdd, the returned matrix is now fill-completed.  Doxygen reflects this choice.
(The underlying EpetraExt method does not do this.)  Switched to using MueLu exceptions.
Replaced a bunch of tabs with spaces.