Remove use of c99 features and move gid type to Zoltan_config.h
Trilinos is c90 compatible so:
Replaced use of %zd with %ld
Removed declarations of variables that occur after instructions in block.
Don't use long long int if LLONG_MAX is not defined.

The type for a ZOLTAN_ID_TYPE is set in cmake this way:

-D Zoltan_ENABLE_UINT_GLOBAL_IDS
-D Zoltan_ENABLE_ULONG_GLOBAL_IDS
-D Zoltan_ENABLE_ULLONG_GLOBAL_IDS

Or in autoconf this way:

--with-id-type=uint
--with-id-type=ulong
--with-id-type=ullong

If the compiler does not fully support unsigned long long the type is changed to
unsigned long at compile time and a warning is printed.

After configuration one of the following is defined in Zoltan_config.h:

and this drives the definition of ZOLTAN_ID_TYPE and related macros in zoltan_types.h.

This fixes bug 5047 because if examples and tests are build with the same Zoltan_config.h
that the library was built with, they will be using the same definition for ZOLTAN_ID_TYPE.