Put ifdef around inclusion of BlockCG when QD enabled (POTRF not available for QD)
When enabling QD support the test in ifpack2/test/belos/belos_extprec_solve.cpp did not compile
because BlockCG requires the LAPACK routine POTRF from Teuchos LAPACK, which is not templated.

I created a new macro USING_QD that is only defined in the above mentioned test program and then
avoid inclusion of BlockCG if this macro is defined.

  On branch master
  Changes to be committed:

 	modified:   belos_extprec_solve.cpp
 	modified:   build_solver.hpp