Fixed error when checked STL is enabled.
While I believe it is a perfectly valid thing to do, the GNU checked-stl
complains if you create a singular iterator (i.e., an iterator that
doesn't point to anything).  This was happening in the new Cijk code
since I create arrays of iterators.  To fix this I supplied an
initializer to the arrays of iterators.  This should fix the errors that
showed up in the nightly tests last night.