re-enable the previously disabled gtest tpl
The gtest tpl was disabled earlier today in response to errors that were caused
by a the commit that introduced it. Gtest was not the cause of the errors, and
changes have been made to STK to no longer require gtest for all builds so it is
safe to re-enable this as a valid tpl. Tests were run to ensure that STK will
configure/build without gtest available.j