Cthulhu/Gallery/MueLu: added helper functions to simplify examples.
For each examples/tests of MueLu, we have to interpret the command line argument to get the parameters of the test case.
By using Teuchos::CommandLineProcessor, we end up with a quite important number of line of code because we have to
- define default parameters
- generate the CommandLineProcessor
- check user inputs
- create a parameter list (== input of gallery::CreateCrsMatrix)

So, I created a class Gallery::Parameters to simplify that. See example/Aggregation/AggregationExample.cpp for an example.

In addition,
cthulhu/src/Def/Cthulhu_Parameters.hpp
allow to switch dynamically between Epetra and Tpetra on examples/tests of MueLu.
It's the same principle as for the gallery.