Updated the LOCA MF support to build against the latest version of MF.
This mostly amounted to passing a new error handler struct into most of
the MF functions, and changing the name of a few others.  There are also
a few new libraries that must be included in the TPL.  It has been a
long time since this was updated, and it was probably hard for people to
get the version that LOCA linked against.